# tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AppConfig": {
        "additionalProperties": false,
        "description": "A managed application: a launch specification, not a window.",
        "properties": {
          "audio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AudioConfig",
                "description": "Audio routing. Absent leaves routing untouched; `{\"output\": null}` silences."
              }
            ]
          },
          "env": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Extra environment variables for the launched process.\n\nApplied last, so they override anything the launcher preset sets.\nHardware acceleration usually needs this: enabling NVDEC on an Nvidia\ncard, for instance, is a matter of `LIBVA_DRIVER_NAME` and\n`NVD_BACKEND` rather than any command-line flag.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "heartbeat": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HeartbeatConfig"
              }
            ]
          },
          "id": {
            "description": "Client-chosen, unique, stable identifier.",
            "type": "string"
          },
          "launcher": {
            "$ref": "#/components/schemas/Launcher"
          },
          "persistProfile": {
            "description": "Keep the browser profile between launches instead of wiping it.",
            "type": "boolean"
          },
          "readiness": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReadinessConfig",
                "description": "Wait for this URL to answer before launching."
              }
            ]
          },
          "restart": {
            "$ref": "#/components/schemas/RestartPolicy"
          }
        },
        "required": [
          "id",
          "launcher"
        ],
        "type": "object"
      },
...
      "BackgroundPreset": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Background"
          },
          {
            "properties": {
              "id": {
                "description": "Client-chosen name, referenced from an output's `background`.",
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "type": "object"
          }
        ],
        "description": "A named background, defined once and used by any number of outputs."
      },
...
      "DesiredState": {
        "additionalProperties": false,
        "description": "The complete desired-state document.",
        "properties": {
          "activeApp": {
            "default": null,
            "description": "Which app is running. `null` runs nothing.\n\nExactly one app is ever active, and it always spans every display —\nthe appliance is a single canvas, not a window manager. Keeping the\nchoice as one pointer makes switching atomic: activating B cannot\nleave A half-enabled the way per-app flags could.",
            "type": [
              "string",
              "null"
            ]
          },
          "apps": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AppConfig"
            },
            "type": "array"
          },
          "backgrounds": {
            "default": [],
            "description": "Named background definitions outputs can refer to.\n\nA multi-display installation usually wants one look across every\nalternative — repeating a wallpaper, scaling mode and color on each\noutput — makes the common case the laborious one and guarantees the\nscreens drift apart the first time somebody edits only three of four.",
            "items": {
              "$ref": "#/components/schemas/BackgroundPreset"
            },
            "type": "array"
          },
          "committed": {
            "default": false,
            "description": "Whether this document is persisted, or a working copy being tried out.\n\nOn reads, Suede reports the truth: `true` for the saved document,\n`false` when a working copy is live. On writes, the *client* speaks:\n`committed: true` persists; anything else applies the document to the\noutputs — reconciled immediately, exactly as if saved — but leaves\ndisk untouched, so a restart or `POST /config/revert` returns to the\nsaved state. A UI can therefore push every edit as it happens and\nonly set the flag when the operator presses Save.",
            "type": "boolean"
          },
          "outputs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OutputConfig"
            },
            "type": "array"
          },
          "projection": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProjectionConfig",
                "description": "Multi-projector features. Absent means no projection processing at all."
              }
            ]
          },
          "revision": {
            "default": 0,
            "description": "Monotonic revision, incremented by Suede on every accepted write.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "schemaVersion": {
            "default": 0,
            "description": "Document schema version, managed by Suede.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "settings": {
            "default": {
              "hideCursor": true,
              "measureCapabilitiesOnStart": true,
              "outputPollIntervalSeconds": 5
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "Settings": {
        "additionalProperties": false,
        "description": "Daemon-level settings that belong to desired state.",
        "properties": {
          "hideCursor": {
            "description": "Hide the pointer and park it beyond the layout.",
            "type": "boolean"
          },
          "measureCapabilitiesOnStart": {
            "description": "Measure browser decode capabilities at startup when the browser,\nits configuration, or the GPU driver changed since last measured.\nA brief window opens on the displays while it runs; with nothing\nchanged, nothing opens.",
            "type": "boolean"
          },
          "outputPollIntervalSeconds": {
            "description": "Backstop poll interval for output changes.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "hideCursor",
          "outputPollIntervalSeconds"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "Optional. When a token is configured, every endpoint except /healthz and the heartbeat endpoint requires it.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Hamish Barjonas"
    },
    "description": "Remote management for Sway-based display appliances. Clients write desired state; a reconciler drives the live session toward it.",
    "license": {
      "name": "PolyForm-Small-Business-1.0.0",
      "url": "https://github.com/gameshowpro/Suede/blob/main/LICENSE"
    },
    "title": "Suede",
    "version": "(set from Cargo.toml)"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesiredState"
                }
              }
            },
            "description": "The current document. `committed: false` means a working copy is live on the outputs but not saved.",
            "headers": {
              "ETag": {
                "description": "Persisted document revision, quoted for If-Match",
                "schema": {
                  "type": "string"
                }
              },
              "X-Config-Epoch": {
                "description": "Store instance identity for If-Config-Epoch",
                "schema": {
                  "type": "string"
                }
              },
              "X-Config-Generation": {
                "description": "Effective working-copy generation",
                "schema": {
                  "format": "int64",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "put_config",
        "parameters": [
          {
            "description": "Block until reconciliation settles, or this many seconds elapse.",
            "in": "query",
            "name": "wait",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Optional persisted revision precondition from ETag",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "Optional working-copy generation precondition",
            "in": "header",
            "name": "If-Config-Generation",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Optional store-instance precondition from X-Config-Epoch",
            "in": "header",
            "name": "If-Config-Epoch",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesiredState"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesiredState"
                }
              }
            },
            "description": "The accepted document",
            "headers": {
              "ETag": {
                "description": "Persisted document revision, quoted for If-Match",
                "schema": {
                  "type": "string"
                }
              },
              "X-Config-Epoch": {
                "description": "Store instance identity for If-Config-Epoch",
                "schema": {
                  "type": "string"
                }
              },
              "X-Config-Generation": {
                "description": "Effective working-copy generation",
                "schema": {
                  "format": "int64",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "If-Match, If-Config-Generation, or If-Config-Epoch is stale"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/v1/config/revert": {
      "post": {
        "operationId": "revert_config",
        "parameters": [
          {
            "description": "Optional persisted revision precondition from ETag",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "Optional working-copy generation precondition",
            "in": "header",
            "name": "If-Config-Generation",
            "required": false,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Optional store-instance precondition from X-Config-Epoch",
            "in": "header",
            "name": "If-Config-Epoch",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesiredState"
                }
              }
            },
            "description": "Working copy discarded; the saved document is re-applied and returned",
            "headers": {
              "ETag": {
                "description": "Persisted document revision, quoted for If-Match",
                "schema": {
                  "type": "string"
                }
              },
              "X-Config-Epoch": {
                "description": "Store instance identity for If-Config-Epoch",
                "schema": {
                  "type": "string"
                }
              },
              "X-Config-Generation": {
                "description": "Effective working-copy generation",
                "schema": {
                  "format": "int64",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "If-Match, If-Config-Generation, or If-Config-Epoch is stale"
          }
        },
        "tags": [
          "config"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A Suede appliance",
      "url": "http://{host}:{port}",
      "variables": {
        "host": {
          "default": "127.0.0.1",
          "description": "Hostname or address of the appliance"
        },
        "port": {
          "default": "9088",
          "description": "The API port"
        }
      }
    }
  ],
...
}